Fixing clippy and fmt issues
