+++ crates/derive/Cargo.toml
@@ -15,3 +15,6 @@ proc-macro = true
 syn = { version = "1.0", features = ["full"] }
 quote = "1.0"
 proc-macro2 = "1.0"
+
+[dev-dependencies]
+kelk = { path = "../kelk" }
